pitch detector finished

+++ app/src/main/java/ru/nobird/junction/model/Vec3f.kt
@@ -4,4 +4,4 @@ package ru.nobird.junction.model
  * Created by Owntage on 11/25/2017.
  */
 
-class Vec3d (val x: Double, val y: Double, val z: Double)
+class Vec3f(val x: Float, val y: Float, val z: Float)
