+++ app/src/main/java/ru/nobird/junction/processing/SensorDataListener.java
@@ -1,12 +1,12 @@
 package ru.nobird.junction.processing;
 
-import ru.nobird.junction.model.Vec3d;
+import ru.nobird.junction.model.Vec3f;
 
 /**
  * Created by Owntage on 11/25/2017.
  */
 
 public interface SensorDataListener {
-    void onLinearAcceleration(Vec3d acceleration, long timestamp);
-    void onAngularAcceleration(Vec3d acceleration, long timestamp);
+    void onLinearAcceleration(Vec3f acceleration, long timestamp);
+    void onAngularAcceleration(Vec3f acceleration, long timestamp);
 }
